Enable eager XNNPACK convolutions for NHWC tensors.
The optimal solution to use XNNPACK is to separate operator creation
from execution - also called prepacking the weights.  If we have done
our job properly, JIT must have caught and replaced nn.Conv2ds on mobile
with the prepacked versions.  Still, if we somehow end up in
_convolution for whatever reason, it is still more efficient to go
through XNNPACK for NHWC tensors, compared to the alternative of
converting NHWC to NCHW and going through NNPACK.

ghstack-source-id: 57ed44be3e45b7a6b07c8baffe1c552c92a175d0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35790